Small modifications

+++ week-01/day 4-5/src/Cuboid.java
@@ -5,7 +5,7 @@ public class Cuboid {
     double c = 15.5;
     double volume = a * b * c;
     double surfaceArea = 2 * (a * b + a * c + b * c);
-    System.out.println("Surface Area: " + surfaceArea);
-    System.out.println("Volume: " + volume);
+    System.out.println("Surface Area: " + (int) surfaceArea);
+    System.out.println("Volume: " + (int) volume);
   }
 }
